+++ app/javascript/utilities/graphics.ts
@@ -363,22 +363,23 @@ export function roundedRectangleHole(
 export function circlePath(loc: Coordinate, r: number): string {
   return [
     "M", loc.x, loc.y-r, "A", r, r, 0, 0, 1, loc.x, loc.y+r,
-    "A", r, r, 0, 0, 1, loc.x, loc.y-r].join(" ")
+    "A", r, r, 0, 0, 1, loc.x, loc.y-r, "z"].join(" ")
 }
 
 export function squarePath(loc: Coordinate): string {
   return [
     "M", loc.x-10, loc.y-10, "L", loc.x+10, loc.y-10, "L", loc.x+10, loc.y+10,
-    "L", loc.x-10, loc.y+10, "L", loc.x-10, loc.y-10].join(" ")
+    "L", loc.x-10, loc.y+10, "z"].join(" ")
 }
 
 export function hexPath(loc: Coordinate, r: number, rotated: boolean): string {
   let a = (rotated ? -0.5 : -1)/3 * Math.PI
   let path = ["M", loc.x + r * Math.cos(a), loc.y + r * Math.sin(a)]
-  for (let i = 0; i < 6; i++) {
+  for (let i = 0; i < 5; i++) {
     a = (i + (rotated ? 0.5 : 0))/3 * Math.PI
     path = path.concat(["L", loc.x + r * Math.cos(a), loc.y + r * Math.sin(a)])
   }
+  path.push("z")
   return path.join(" ")
 }